chore: update to v0.1.2 and rename eth_abis
