+++ bouncer/shared/contract_interfaces.ts
@@ -15,7 +15,7 @@ function loadContractCached(abiPath: string) {
   };
 }
 const CF_ETH_CONTRACT_ABI_TAG = 'v1.0.0';
-const CF_SOL_PROGRAM_IDL_TAG = 'v0.1.1';
+const CF_SOL_PROGRAM_IDL_TAG = 'v0.1.2';
 export const getErc20abi = loadContractCached(
   '../contract-interfaces/eth-contract-abis/IERC20.json',
 );
