+++ bouncer/shared/send_erc20.ts
@@ -2,7 +2,7 @@ import Web3 from 'web3';
 import { Chain } from '@chainflip-io/cli/.';
 import { signAndSendTxEvm } from './send_evm';
 import { amountToFineAmount, getEvmEndpoint } from './utils';
-import { getErc20abi } from './eth_abis';
+import { getErc20abi } from './contract_interfaces';
 
 const erc20abi = await getErc20abi();
