+++ localnet/docker-compose.yml
@@ -9,7 +9,7 @@ services:
     command: /bin/sh -c "cp -R /initial-state/* /localnet-initial-state"
 
   init-solana:
-    image: ghcr.io/chainflip-io/solana-localnet-ledger:v0.1.1
+    image: ghcr.io/chainflip-io/solana-localnet-ledger:v0.1.2
     container_name: init-solana
     platform: linux/amd64
     volumes:
